fix: add lobby service check for lobby availability

+++ back/Services/Lobbies/ILobbyConductService.cs
@@ -15,7 +15,7 @@ namespace Quizer.Services.Lobbies
         public Result<LobbyStatus> GetLobbyStatus(string lobbyGuid);
         public Result<QuestionData> GetCurrentQuestion(string lobbyGuid);
         public Task<Result> RegisterTestAnswer(string userId, string lobbyGuid, string? answerGuid);
-        public Task<Result> RegisterNumericalAnswer(string userId, string lobbyGuid, int answer);
+        public Task<Result> RegisterNumericalAnswer(string userId, string lobbyGuid, float answer);
         public Task<Result> RegisterTextAnswer(string userId, string lobbyGuid, string answer);
         public Result<IEnumerable<AnswerData>> GetRightAnswers(string lobbyGuid);
     }
